Implemented chapter 7 of the book. Adding a world, a view transformation and a camera. Refactored the Amatrix4 class by removing all classes deriving from it so that all transformation matrices are now methods and re-organised all unit tests into their corresponding chapters instead of by each class type.

+++ RayTracerChallenge/ConRayTracer/src/vector.cpp
@@ -43,8 +43,8 @@ Vector Vector::operator/(const float& scalar) const {
                              z_ / scalar);
 }
 
-double Vector::magnitude() const {
-  return sqrt(pow(x_, 2) + pow(y_, 2) + pow(z_, 2));
+float Vector::magnitude() const {
+  return (float)sqrt(pow(x_, 2.0f) + pow(y_, 2.0f) + pow(z_, 2.0f));
 }
 
 Vector Vector::normalize() const {
